#2517 move more functions to API namespace
