+++ src/components/widget/Attributes/Attributes.js
@@ -2,8 +2,12 @@ import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import classnames from 'classnames';
 
-import { getAttributesInstance, initLayout } from '../../../api';
-import { completeRequest, patchRequest } from '../../../actions/GenericActions';
+import {
+  getAttributesInstance,
+  initLayout,
+  patchRequest,
+} from '../../../api';
+import { completeRequest } from '../../../actions/GenericActions';
 import { parseToDisplay } from '../../../utils/documentListHelper';
 import AttributesDropdown from './AttributesDropdown';
