+++ src/containers/Board.js
@@ -5,14 +5,9 @@ import { DragDropContext } from 'react-dnd';
 import HTML5Backend from 'react-dnd-html5-backend';
 import { connect } from 'react-redux';
 
+import { getData, patchRequest } from '../api';
 import { addCard } from '../actions/BoardActions';
-
-import {
-  deleteRequest,
-  getData,
-  getRequest,
-  patchRequest,
-} from '../actions/GenericActions';
+import { deleteRequest, getRequest } from '../actions/GenericActions';
 import { connectWS, disconnectWS } from '../utils/websockets';
 
 import BlankPage from '../components/BlankPage';
